@@ -1,6 +1,6 @@
 {
   "name": "creativefactoryrv/google-maps-geocoder",
-  "version": "1.0.5",
+  "version": "2.0.0",
   "description": "Simple and lightweight geocoder that uses the Google Maps Platform API. It can optionally use a PSR cache.",
   "keywords": [
     "Google Maps Platform",
@@ -13,8 +13,8 @@
   "type": "library",
   "license": "MIT",
   "require": {
-    "php": ">=7.0",
-    "psr/cache": "^1.0"
+    "php": ">=8.0",
+    "psr/cache": "^2.0"
   },
   "autoload": {
     "psr-4": {
